feat(lambda): implement tag and permission operations with storage persistence
